% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman.R
\name{kalman}
\alias{kalman}
\title{kalman filtering and smoothing for vector autoregression with measurement error}
\usage{
kalman(Y,A,sig_eta,sig_epsilon,X_init=NULL,P_init=NULL)
}
\arguments{
\item{Y}{observations of time series, a p by T matrix.}

\item{A}{current estimate of transition matrix.}

\item{sig_eta}{current estiamte of \eqn{\sigma_\eta}.}

\item{sig_epsilon}{current estiamte \eqn{\sigma_\epsilon}.}

\item{X_init}{inital estimate of latent \eqn{x_1} at the first iteration, a p-dimensional vector.}

\item{P_init}{inital covariance estimate of latent \eqn{x_1} at the first iteration, a p by p matrix.}
}
\value{
a list of conditional expectations and covariances of \eqn{x_t}'s.
}
\description{
kalman filtering and smoothing for vector autoregression with measurement error
}
\author{
Xiang Lyu, Jian Kang, Lexin Li
}
