% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARMLE.R
\name{VARMLE}
\alias{VARMLE}
\title{generalized Dantzig selector for transition matrix update in maximization step}
\usage{
VARMLE(S0, S1, tol)
}
\arguments{
\item{S0}{a p by p matrix; average (over time points) of conditional expectation of \eqn{x_t x_t^\top} on \eqn{y_1, \ldots, y_T} and parameter estimates, obtained from expectation step.}

\item{S1}{a p by p matrix; average (over time points) of conditional expectation of \eqn{x_t x_{t+1}^\top}on \eqn{y_1, \ldots, y_T} and parameter estimates, obtained from expectation step.}

\item{tol}{tolerance parameter in Dantzig selector.}
}
\value{
Sparse estimate of transition matrix by Dantzig selector.
}
\description{
Sparse estimation of transtion matrix in vector autoregression given conditional autocovariance matrices.
}
\author{
Xiang Lyu, Jian Kang, Lexin Li
}
